Move some classes to types.py Update MBFingerprinterPlugin name field
