+++ mporg/plugins/plugin_loader.py
@@ -2,6 +2,7 @@ import importlib
 import inspect
 import os
 import sys
+from pathlib import Path
 
 from mporg.plugins import FINGERPRINTER_DIR, PLUGIN_DIR
 from mporg.plugins.util import PluginType, Plugin
@@ -19,9 +20,9 @@ class PluginLoader:
         self.cred_managers = {}
         self.fingerprinters = {}
 
-    def fingerprinter_from_file(self, file: str):
-        plugin_name = file.replace("Plugin.py", "")
-        module_name = plugin_name + "Plugin"
+    def fingerprinter_from_file(self, file: Path):
+        module_name = file.stem
+        plugin_name = module_name.replace("Plugin", "")
 
         try:
             spec = importlib.util.spec_from_file_location(module_name, FINGERPRINTER_DIR / file)
@@ -41,17 +42,19 @@ class PluginLoader:
         if FINGERPRINTER_DIR.exists():
             print(FINGERPRINTER_DIR)
 
-        fingerprinter_files = [f for f in os.listdir(PLUGIN_DIR / plugin_type.value) if f.endswith("Plugin.py")]
-        fil_name = plugin_name.lower().strip()
-        file = [f for f in fingerprinter_files if fil_name in f.lower().strip()][0]
+        files = [f for f in (PLUGIN_DIR / plugin_type.value).rglob("*Plugin.py")]
 
-        self.fingerprinter_from_file(file)
+        fil_name = plugin_name.lower().strip()
+        file = [f for f in files if fil_name in str(f).lower().strip()][0]
+
+        if plugin_type == PluginType.FINGERPRINTER:
+            self.fingerprinter_from_file(file)
 
     def load_all_fingerprinters(self):
         if FINGERPRINTER_DIR.exists():
             print(FINGERPRINTER_DIR)
 
-        fingerprinter_files = [f for f in os.listdir(FINGERPRINTER_DIR) if f.endswith("Plugin.py")]
+        fingerprinter_files = list(FINGERPRINTER_DIR.rglob("*Plugin.py"))
         for file in fingerprinter_files:
             self.fingerprinter_from_file(file)
