+++ tests/test_org.py
@@ -10,6 +10,7 @@ from parameterized import parameterized
 
 import mporg.main
 import mporg.organizer as mp
+import mporg.types
 from mporg.logging_utils.logging_setup import setup_logging
 
 from tests import utils
@@ -46,28 +47,28 @@ class TestMPORGHelpers(unittest.TestCase):
 
     @parameterized.expand([
         ("Within Limits",
-         mp.Track(album_artists=("Artist1", "Artist2",),
-                  track_artists=("Artist3",),
-                  album_name="Album Name",
-                  track_name="Track Name"),
+         mporg.types.Track(album_artists=("Artist1", "Artist2",),
+                           track_artists=("Artist3",),
+                           album_name="Album Name",
+                           track_name="Track Name"),
          ("Artist1, Artist2", "Album Name", "Artist3", "Track Name")),
         ("Long Album Name",
-         mp.Track(album_artists=("Artist1", "Artist2",),
-                  track_artists=("Artist3",),
-                  album_name=long_message + "This is above",
-                  track_name="Track Name"),
+         mporg.types.Track(album_artists=("Artist1", "Artist2",),
+                           track_artists=("Artist3",),
+                           album_name=long_message + "This is above",
+                           track_name="Track Name"),
          ("Artist1, Artist2", 'x' * name_max, "Artist3", "Track Name")),
         ("Long Track Name",
-         mp.Track(album_artists=("Artist1", "Artist2",),
-                  track_artists=("Artist3",),
-                  album_name="Album Name",
-                  track_name=long_message + "This is above"),
+         mporg.types.Track(album_artists=("Artist1", "Artist2",),
+                           track_artists=("Artist3",),
+                           album_name="Album Name",
+                           track_name=long_message + "This is above"),
          ("Artist1, Artist2", "Album Name", "Artist3", 'x' * name_max)),
         ("Long Track and Album",
-         mp.Track(album_artists=tuple(["Artist1" * path_max]),  # Much larger thna path max
-                  track_artists=("Artist2" * path_max,),
-                  album_name=long_message,  # 10 characters
-                  track_name=long_message),  # 10 characters
+         mporg.types.Track(album_artists=tuple(["Artist1" * path_max]),  # Much larger thna path max
+                           track_artists=("Artist2" * path_max,),
+                           album_name=long_message,  # 10 characters
+                           track_name=long_message),  # 10 characters
          (("Artist1" * 24)[:artist_max],  # 30 characters
           "x" * name_max,  # 10 characters
           ("Artist2" * 20)[:artist_max],  # 30 characters
@@ -90,14 +91,14 @@ class TestMPORG(unittest.TestCase):
 
     def test_get_metadata_with_spotify_results(self):
         metadata = utils.MockTagger(Path("song.mp3"), {'title': ["Song 1"], "artist": ["Artist 1"]})
-        self.org.search_spotify = MagicMock(return_value=mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.org.search_spotify = MagicMock(return_value=mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
         self.org.get_fingerprint_metadata = MagicMock()
         self.org.get_fingerprint_spotify_metadata = MagicMock()
 
         file = MagicMock()
         results, source = self.org.get_metadata(metadata, file)
 
-        self.assertEqual(results, mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.assertEqual(results, mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
         self.assertEqual(source, mp.TagType.SPOTIFY)
         self.org.search_spotify.assert_called_once_with('Song 1', 'Artist 1')
         self.org.get_fingerprint_metadata.assert_not_called()
@@ -107,27 +108,27 @@ class TestMPORG(unittest.TestCase):
         metadata = utils.MockTagger(Path("song.mp3"), {'title': ["Song 1"], "artist": ["Artist 1"]})
         self.org.search_spotify = MagicMock(return_value=None)
         self.org.get_fingerprint_metadata = MagicMock(
-            return_value=mp.FingerprintResult(0, 'suc', mp.Track(track_name="Song 1", track_artists=("Artist 1",))))
+            return_value=mp.FingerprintResult(0, 'suc', mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",))))
         self.org.get_fingerprint_spotify_metadata = MagicMock()
 
         file = MagicMock()
         results, source = self.org.get_metadata(metadata, file)
 
         self.assertEqual(results,
-                         mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+                         mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
         self.assertEqual(source, mp.TagType.FINGERPRINTER)
         self.org.search_spotify.assert_called_once_with('Song 1', 'Artist 1')
         self.org.get_fingerprint_metadata.assert_called_once_with(file)
         self.org.get_fingerprint_spotify_metadata.assert_not_called()
 
     def test_search_spotify(self):
-        self.org.sh.search = MagicMock(return_value=mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.org.sh.search = MagicMock(return_value=mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
 
         title = 'Song 1'
         artist = 'Artist 1'
         results = self.org.search_spotify(title, artist)
 
-        self.assertEqual(results, mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.assertEqual(results, mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
         self.org.sh.search.assert_called_once_with(name='Song 1', artist='Artist 1')
 
     def test_search_spotify_without_results(self):
@@ -166,14 +167,14 @@ class TestMPORG(unittest.TestCase):
                                                                                 type="spotify",
                                                                                 results={'spotifyid': 12345}))
         sh = self.org.sh = MagicMock()
-        sh.search.side_effect = [None, mp.Track(track_name="Song 1", track_artists=("Artist 1",))]
+        sh.search.side_effect = [None, mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",))]
 
         tag = utils.MockTagger(Path("song.mp3"), {'title': ["Song 1"], "artist": ["Artist 1"]})
 
         file = MagicMock()
         results, tagType = self.org.get_metadata(tag, file)
 
-        self.assertEqual(results, mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.assertEqual(results, mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
         self.assertEqual(tagType, mp.TagType.SPOTIFY)
         fingerprinter.fingerprint.assert_called_once_with(file)
 
@@ -181,8 +182,8 @@ class TestMPORG(unittest.TestCase):
         fingerprinter = self.org.af[0] = MagicMock()
         fingerprinter.fingerprint = MagicMock(return_value=mp.FingerprintResult(code=0,
                                                                                 type="fingerprinter",
-                                                                                results=mp.Track(track_name="Song 1",
-                                                                                                 track_artists=("Artist 1",))))
+                                                                                results=mporg.types.Track(track_name="Song 1",
+                                                                                                          track_artists=("Artist 1",))))
         sh = self.org.sh = MagicMock()
         sh.search.return_value = None
 
@@ -191,7 +192,7 @@ class TestMPORG(unittest.TestCase):
         file = MagicMock()
         results, tagType = self.org.get_metadata(tag, file)
 
-        self.assertEqual(results, mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.assertEqual(results, mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
         self.assertEqual(tagType, mp.TagType.FINGERPRINTER)
         fingerprinter.fingerprint.assert_called_once_with(file)
 
@@ -210,12 +211,12 @@ class TestMPORG(unittest.TestCase):
         self.assertEqual(tagType, mp.TagType.METADATA)
 
     def test_get_fingerprint_spotify_metadata(self):
-        self.org.sh.search = MagicMock(return_value=mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.org.sh.search = MagicMock(return_value=mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
 
         spotify_id = '12345'
         results = self.org.get_fingerprint_spotify_metadata(spotify_id)
 
-        self.assertEqual(results, mp.Track(track_name="Song 1", track_artists=("Artist 1",)))
+        self.assertEqual(results, mporg.types.Track(track_name="Song 1", track_artists=("Artist 1",)))
         self.org.sh.search.assert_called_once_with(spot_id='12345')
 
     def test_get_fingerprint_spotify_metadata_without_results(self):
@@ -230,9 +231,9 @@ class TestMPORG(unittest.TestCase):
     def test_get_location_spotify(self):
         # Test get_location function with TagType.SPOTIFY
 
-        results = mp.Track(album_artists=("Drag",), album_name="mporgTests",
-                           track_artists=("Drag",), track_name="spotTest",
-                           track_number=1, album_year="2023")
+        results = mporg.types.Track(album_artists=("Drag",), album_name="mporgTests",
+                                    track_artists=("Drag",), track_name="spotTest",
+                                    track_number=1, album_year="2023")
 
         tags_from = mp.TagType.SPOTIFY
         metadata = utils.MockTagger(Path("song.mp3"))
@@ -243,12 +244,12 @@ class TestMPORG(unittest.TestCase):
         self.assertEqual(actual_path, expected_path)
 
     @parameterized.expand([
-        ("All Data", mp.Track(album_artists=("Drag",), album_name="mporgTests",
-                              track_artists=("Drag",), track_name="fingerTest",
-                              track_number=1, track_year="2023"),
+        ("All Data", mporg.types.Track(album_artists=("Drag",), album_name="mporgTests",
+                                       track_artists=("Drag",), track_name="fingerTest",
+                                       track_number=1, track_year="2023"),
          Path("store/Drag/2023 - mporgTests/Drag - fingerTest.mp3")),
-        ("No Track Year", mp.Track(album_artists=("Drag",), album_name="mporgTests",
-                                   track_artists=("Drag",), track_name="fingerTest"),
+        ("No Track Year", mporg.types.Track(album_artists=("Drag",), album_name="mporgTests",
+                                            track_artists=("Drag",), track_name="fingerTest"),
          Path("store/Drag/mporgTests/Drag - fingerTest.mp3"))
     ])
     def test_get_location_with_fingerprinter_tags(self, name, results, expected_path):
@@ -273,7 +274,7 @@ class TestMPORG(unittest.TestCase):
     ])
     def test_get_location_with_metadata_tags(self, name, metadata, expected_path):
         """Test get_location function with TagType.METADATA"""
-        results = mp.Track()
+        results = mporg.types.Track()
         tags_from = mp.TagType.METADATA
         dummy_tagger = utils.MockTagger(Path("song.mp3"), metadata)
         file = Path("song.mp3")
@@ -295,10 +296,10 @@ class TestMPORG(unittest.TestCase):
     @patch('mporg.organizer.Tagger')
     def test_update_metadata_from_spotify(self, mock_tagger):
         location = Path('song.mp3')
-        results = mp.Track(track_name='Test Track', track_artists=('Test Artist',), album_name='Test Album',
-                           album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
-                           album_artists=('Test Artist',), track_bpm='120', track_key='C',
-                           album_genres='Rock')
+        results = mporg.types.Track(track_name='Test Track', track_artists=('Test Artist',), album_name='Test Album',
+                                    album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
+                                    album_artists=('Test Artist',), track_bpm='120', track_key='C',
+                                    album_genres='Rock')
 
         self.org.update_metadata_from_spotify(threading.Lock(), location, results)
 
@@ -324,10 +325,10 @@ class TestMPORG(unittest.TestCase):
     @patch('mporg.organizer.Tagger')
     def test_update_metadata_from_fingerprinter(self, mock_tagger):
         location = Path('song.mp3')
-        results = mp.Track(track_name='Test Track', track_artists=('Test Artist',), album_name='Test Album',
-                           album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
-                           album_artists=('Test Artist',), track_bpm='120', track_key='C',
-                           album_genres='Rock', track_year='2023')
+        results = mporg.types.Track(track_name='Test Track', track_artists=('Test Artist',), album_name='Test Album',
+                                    album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
+                                    album_artists=('Test Artist',), track_bpm='120', track_key='C',
+                                    album_genres='Rock', track_year='2023')
 
         self.org.update_metadata_from_fingerprinter(threading.Lock(), location, results)
 
@@ -370,10 +371,10 @@ class TestMPORGWithMocks(unittest.TestCase):
 
     @patch('mporg.organizer.Tagger')
     def test_process_file_spotify(self, tag):
-        spotifyRes = mp.Track(track_name='Test Track', track_artists=('Test Artist',), album_name='Test Album',
-                              album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
-                              album_artists=('Test Artist',), track_bpm='120', track_key='C',
-                              album_genres='Rock')
+        spotifyRes = mporg.types.Track(track_name='Test Track', track_artists=('Test Artist',), album_name='Test Album',
+                                       album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
+                                       album_artists=('Test Artist',), track_bpm='120', track_key='C',
+                                       album_genres='Rock')
         # Set up expected return values and mocks
         self.mporg.get_metadata = MagicMock(
             return_value=(spotifyRes, mp.TagType.SPOTIFY)
