+++ tests/utils.py
@@ -1,10 +1,11 @@
 from concurrent.futures import Future, ThreadPoolExecutor
 from unittest.mock import MagicMock
 
+import mporg.types
 from mporg import organizer, spotify_searcher, audio_fingerprinter
 
 
-class MockTagger(organizer.Tagger):
+class MockTagger(mporg.types.Tagger):
     def __init__(self, file, fake=None):
         if fake is None:
             fake = dict()
@@ -29,20 +30,20 @@ class MockTagger(organizer.Tagger):
 
 class MockSearcher:
     def search(self, name: str = None, artist: str = None, spot_id: str = None):
-        return spotify_searcher.Track(track_name='Test Track', track_artists=['Test Artist'], album_name='Test Album',
-                                      album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
-                                      album_artists=['Test Artist'], track_bpm='120', track_key='C',
-                                      album_genres='Rock')
+        return mporg.types.Track(track_name='Test Track', track_artists=['Test Artist'], album_name='Test Album',
+                                 album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
+                                 album_artists=['Test Artist'], track_bpm='120', track_key='C',
+                                 album_genres='Rock')
 
 
 class MockFingerprinter(audio_fingerprinter.Fingerprinter):
     def fingerprint(self, path):
         return audio_fingerprinter.FingerprintResult(
             0, "spotify",
-            spotify_searcher.Track(track_name='Test Track', track_artists=['Test Artist'], album_name='Test Album',
-                                   album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
-                                   album_artists=['Test Artist'], track_bpm='120', track_key='C',
-                                   album_genres='Rock')
+            mporg.types.Track(track_name='Test Track', track_artists=['Test Artist'], album_name='Test Album',
+                              album_year='2023', track_number=1, track_disk=1, track_url='http://example.com',
+                              album_artists=['Test Artist'], track_bpm='120', track_key='C',
+                              album_genres='Rock')
         )
